#32: Refactor backend to hexagonal architecture
## Architecture Changes
- Restructure backend to follow hexagonal/clean architecture principles
- Separate concerns into core, application, and infrastructure layers

## Core Layer (Domain)
- Move all domain models to core/filesystem/model package
- Consolidate User, File, Directory, Project, and related models
- Keep domain logic independent of frameworks

## Application Layer (Use Cases)
- Create service layer in application/filesystem/service
- Add AuthService, UserService, FileService, ProjectService
- Move OidcUserSyncService to application layer
- Implement business logic and orchestration

## Infrastructure Layer (Adapters)
- Move REST controllers to infrastructure/filesystem/in/web/rest
- Relocate FileController, ProjectController, AuthController, UserController
- Controllers now delegate to application services

## Repository Updates
- Update all repository imports to use new core model locations
- Update savers to reference new model packages
- Maintain existing repository functionality

## Test Updates
- Update test imports to match new package structure
- Ensure all 33 tests continue to pass
- Update FileControllerTest and ProjectControllerTest

This refactoring improves maintainability, testability, and follows clean architecture best practices while maintaining all existing functionality.
