+++ backend/src/main/java/edu/kit/quak/core/filesystem/model/UserDto.java
@@ -1,4 +1,4 @@
-package edu.kit.quak.security.model;
+package edu.kit.quak.core.filesystem.model;
 
 import java.util.UUID;
