+++ backend/src/main/java/edu/kit/quak/files/configuration/FileElementResolver.java
@@ -3,9 +3,9 @@ package edu.kit.quak.files.configuration;
 import com.fasterxml.jackson.databind.DatabindContext;
 import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.jsontype.impl.TypeIdResolverBase;
-import edu.kit.quak.files.model.Directory;
-import edu.kit.quak.files.model.File;
-import edu.kit.quak.files.model.FileElement;
+import edu.kit.quak.core.filesystem.model.Directory;
+import edu.kit.quak.core.filesystem.model.File;
+import edu.kit.quak.core.filesystem.model.FileElement;
 import edu.kit.quak.files.repository.savers.FileElementSaver;
 import edu.kit.quak.files.repository.savers.FileElementSaversRepository;
 import org.springframework.beans.factory.annotation.Autowired;
