+++ backend/src/main/java/edu/kit/quak/files/repository/DirectoryRepository.java
@@ -1,6 +1,6 @@
 package edu.kit.quak.files.repository;
 
-import edu.kit.quak.files.model.Directory;
+import edu.kit.quak.core.filesystem.model.Directory;
 import org.springframework.data.repository.CrudRepository;
 
 public interface DirectoryRepository extends CrudRepository<Directory, String> {
