+++ backend/src/main/java/edu/kit/quak/files/repository/FileRepository.java
@@ -1,6 +1,6 @@
 package edu.kit.quak.files.repository;
 
-import edu.kit.quak.files.model.File;
+import edu.kit.quak.core.filesystem.model.File;
 import org.springframework.data.repository.CrudRepository;
 
 public interface FileRepository extends CrudRepository<File, String> {
