+++ backend/src/main/java/edu/kit/quak/files/repository/ProjectRepository.java
@@ -1,9 +1,9 @@
 package edu.kit.quak.files.repository;
 
-import edu.kit.quak.files.model.Project;
+import edu.kit.quak.core.filesystem.model.Project;
 import org.springframework.data.repository.CrudRepository;
 
-import edu.kit.quak.security.model.User;
+import edu.kit.quak.core.filesystem.model.User;
 import java.util.List;
 
 public interface ProjectRepository extends CrudRepository<Project, String> {
