+++ backend/src/main/java/edu/kit/quak/security/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package edu.kit.quak.security.repository;
 
-import edu.kit.quak.security.model.User;
+import edu.kit.quak.core.filesystem.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
